fwpm_backend/apps/network_performance/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
pandas>=2.0.0
trino>=0.327.0
django-redis>=5.4.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
python-dateutil>=2.8.2
certifi>=2024.2.2
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
django-filter>=23.5
django-cors-headers>=4.3.1 